Fix snow bug
The formula in lines 1161-1162 is calculating a volume increase rather than a height increase, as expected. As a result, an excessive increase in height is calculated, bringing water level and volume outside the range of values of the provided bathymetry and making the model to crash, sometimes with the error "Array bounds error - too many layers" (e.g., see https://groups.google.com/forum/#!category-topic/aquaticmodelling/glm/2ZxGiKXI5Fc). The suggested modification fixes this issue.